remove cereal port dependency. improved error handling.

+++ cyberglove/include/cyberglove/cyberglove_publisher.h
@@ -67,7 +67,7 @@ private:
   //ros node handle
   NodeHandle node, n_tilde;
   Rate sampling_rate;
-  unsigned int publish_counter_max, publish_counter_index;
+  size_t publish_counter_max, publish_counter_index;
 
   ///the actual connection with the cyberglove is done here.
   boost::scoped_ptr<CybergloveSerial> serial_glove;
